fix wrapper test
